Added tracking of working time and other necessary attributes. Added softFilter preferentially assignes assistants that didn't recently have shift.

+++ src/main/java/graphics/sorter/Filters/AssistantMonthShift.java
@@ -8,7 +8,7 @@ public class AssistantMonthShift {
 
 
     private UUID assistantID;
-    private double workedHours;
+    private long workedHours;
     private int lastDayInWork;
 
     public AssistantMonthShift(UUID id){
@@ -23,11 +23,11 @@ public class AssistantMonthShift {
         this.assistantID = assistantID;
     }
 
-    public double getWorkedHours() {
+    public long getWorkedHours() {
         return workedHours;
     }
 
-    public void setWorkedHours(double workedHours) {
+    public void setWorkedHours(long workedHours) {
         this.workedHours = workedHours;
     }
